// src/app/interfaces/movies-list.ts
export interface MoviesList {
  id?: number;
  title?: string;
  usGross?: number;
  usDvdSales?: number;
  worldWideGross?: number;
  productionBudget?: number;
  releaseDate?: string;
  distributor?: string;
  imdbRating?: number;
  imdbVotes?: number;
  majorGenre?: string;
  director?: string;
  rottenTomatoesRating?: string;
  mpaaRating?: string;
  source?: string;
  runningTimeMin?: string;
  creativeType?: string;
  sort?: string;
  sortValue?: string;
}